massive code cleanup!

+++ gdx/src/com/badlogic/gdx/maps/tiled/renderers/HexagonalTiledMapRenderer.java
@@ -16,8 +16,6 @@
 
 package com.badlogic.gdx.maps.tiled.renderers;
 
-import static com.badlogic.gdx.graphics.g2d.Batch.*;
-
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
@@ -26,6 +24,8 @@ import com.badlogic.gdx.maps.tiled.TiledMapImageLayer;
 import com.badlogic.gdx.maps.tiled.TiledMapTile;
 import com.badlogic.gdx.maps.tiled.TiledMapTileLayer;
 import com.badlogic.gdx.maps.tiled.tiles.AnimatedTiledMapTile;
+
+import static com.badlogic.gdx.graphics.g2d.Batch.*;
 
 public class HexagonalTiledMapRenderer extends BatchTiledMapRenderer {
 
@@ -84,12 +84,12 @@ public class HexagonalTiledMapRenderer extends BatchTiledMapRenderer {
 
 		Integer length = map.getProperties().get("hexsidelength", Integer.class);
 		if (length != null) {
-			hexSideLength = length.intValue();
+			hexSideLength = length;
 		} else {
 			if (staggerAxisX) {
 				length = map.getProperties().get("tilewidth", Integer.class);
 				if (length != null) {
-					hexSideLength = 0.5f * length.intValue();
+					hexSideLength = 0.5f * length;
 				} else {
 					TiledMapTileLayer tmtl = (TiledMapTileLayer)map.getLayers().get(0);
 					hexSideLength = 0.5f * tmtl.getTileWidth();
@@ -97,7 +97,7 @@ public class HexagonalTiledMapRenderer extends BatchTiledMapRenderer {
 			} else {
 				length = map.getProperties().get("tileheight", Integer.class);
 				if (length != null) {
-					hexSideLength = 0.5f * length.intValue();
+					hexSideLength = 0.5f * length;
 				} else {
 					TiledMapTileLayer tmtl = (TiledMapTileLayer)map.getLayers().get(0);
 					hexSideLength = 0.5f * tmtl.getTileHeight();
